requirements.txt
flask>=2.0.0
pandas>=1.5.0
pyodbc>=4.0.0
requests>=2.28.0
python-dotenv>=1.0.0
pyarrow>=10.0.0
openpyxl>=3.0.0
werkzeug>=2.0.0